moved the menu prompts back into the index.js file
