+++ lib/index.js
@@ -1,8 +1,5 @@
 const inquirer = require('inquirer');
-const getDepartments = require('./lib/getDepartments.js');
-const viewAllRoles = require('./lib/viewAllRoles.js');
-const addNewDepartment = require('./lib/addNewDepartment.js');
-
+const getDepartments = require('./lib/get.departments.js');
 
 
 const employeeManagerStr = `
@@ -23,22 +20,18 @@ const employeeManagerStr = `
                                                                      
 
 `;
+
 console.log(employeeManagerStr);
 
-
-
-const managerMenuItems = [
+const questions = [
     {
         type: 'list',
         name: 'action',
         message: 'Select an action:',
         choices: [
-            new inquirer.Separator(),
-            'Exit Employee Manager',
-            new inquirer.Separator(),
             'View All Departments',
             'View All Employees',
-            'View All Roles',
+            'View All Roles', 
             'Add a Department',
             'Add an Employee',
             'Add a Role',
@@ -48,66 +41,57 @@ const managerMenuItems = [
             // 'View Employees by Department', //Bonus
             // 'Delete Departments, Roles, and Employees', // Bonus
             // 'View Utilized Budget of a Department',  // Bonus: i.e., the combined salaries of all employees in that department
-            // new inquirer.Separator(),
-            // 'Exit Employee Manager',
-            // new inquirer.Separator(),
+            new inquirer.Separator(),
+            'Exit Employee Manager',
+            new inquirer.Separator(),
         ],
     },
 ];
 
+let isManagerActive = true;
+async function main() {
+    while (isManagerActive) {
+        console.log('Main Menu');
+        const answers = await inquirer.prompt(questions);
 
-async function managerRouter(action) {
-    
-    switch (action) {
-
-        case 'Exit Employee Manager':
+        if (answers.action === 'Exit Employee Manager') {
             console.log('Goodbye!');
-            return;
+            isManagerActive = false;
+            break; // Exit the loop when the user selects "Exit Employee Manager"
+        }
 
-        case 'View All Departments':
-            var response = await getDepartments();
-            (response.status === "success") ? console.table(response.body) : console.log(response);
-            break;
-
-        case 'View All Roles':
-            viewAllRoles();
-            // managerMenu();
-            break;
-
-        case 'View All Employees':
-            break;
-
-        case 'Add a Department':
-            var response = await addNewDepartment();
-            if(response !== "success") console.log(response);
-            break;
-
-        case 'Add a Role':
-            break;
-
-        case 'Add an Employee':
-            break;
-
-        case 'Update an Employee Role':
-            break;
-
-        default:
-            console.log('Error: Menu Action functionality does not exist. Please report error to your software development team.');
+        // Handle other options here based on answers.action
+        switch (answers.action) {
+            case 'View All Departments':
+                setTimeout(() => {
+                    console.log('Viewing All Departments:');
+                    getDepartments();
+                    console.log("does it get here?");
+                }, 1000);
+                break;
+            case 'View All Roles':
+                console.log('Viewing All Roles:');
+                break;
+            case 'View All Employees':
+                console.log('Viewing All Employees:');
+                break;
+            case 'Add a Department':
+                console.log('Adding a Department:');
+                break;
+            case 'Add a Role':
+                console.log('Adding a Roles:');
+                break;
+            case 'Add an Employee':
+                console.log('Adding an Employee:');
+                break;
+            case 'Update an Employee Role':
+                console.log('Updating an Employee Role:');
+                break;
+            
+        }
     }
-
-    // managerMenu();
-    setTimeout(() => {
-        managerMenu();
-    }, 1000);
 }
 
+main();
 
-
-function managerMenu() {
-
-    console.log('\nMain Menu\n');
-    inquirer.prompt(managerMenuItems).then((selected) => { managerRouter(selected.action); });
-
-}
-managerMenu();
-
+console.log("Does it make it past the main() call??");
