+++ lib/logo-index.js
@@ -1,5 +1,5 @@
 // Install Dependencies
-const Logo = require('./lib/Logo.js');
+const Logo = require('./Logo.js');
 
 let employeeManagerPrompts = true;
