+++ lib/manager-index.js
@@ -1,4 +1,4 @@
-const managerMenu = require('./lib/manager.menu.js');
+const managerMenu = require('./manager.menu.js');
 
 const employeeManagerStr = `
